+++ src/app/lefmarna-otoiawase/thanks/page.tsx
@@ -1,27 +1,10 @@
-'use client'
+import { Metadata } from 'next'
+import { LefmarnaOtoiawaseThanks } from './LefmarnaOtoiawaseThanks'
 
-import { Container, Button } from '@mui/material'
-// import Title from '@/components/Title'
-import { useRouter } from 'next/navigation'
+export const metadata: Metadata = {
+  title: '送信成功',
+}
 
-export default function Thanks() {
-  const router = useRouter()
-
-  return (
-    <Container>
-      {/* <p text="チップを贈る" /> */}
-      <p>
-        応援ありがとうございます。
-        <br />
-        引き続き『Pokemonote』をよろしくお願いします。
-      </p>
-      <Button
-        onClick={() => router.push('/')}
-        variant="contained"
-        color="primary"
-      >
-        トップページへ戻る
-      </Button>
-    </Container>
-  )
+export default function Page() {
+  return <LefmarnaOtoiawaseThanks />
 }
